refactor: simplify animation methods
